+++ react/test-setup.js
@@ -6,6 +6,11 @@ const jsdom = new JSDOM('<!doctype html><html><body></body></html>')
 
 global.window = jsdom.window
 global.document = window.document
+global.__RUNTIME__ = {
+  pages: {
+    'store/search': '/path',
+  },
+}
 
 Enzyme.configure({ adapter: new Adapter() })